mode_declares

+++ code/controllers/configuration.dm
@@ -833,7 +833,7 @@ var/global/bridge_secret = null
 		if(probabilities[M.config_name] <= 0)
 			qdel(M)
 			continue
-		if(global.master_last_mode == M.name)
+		if(global.master_last_mode == M.name && M.name != "Extended")
 			qdel(M)
 			continue
 		if(global.modes_failed_start[M.name])
